+++ tests/Deprecated.php
@@ -1,0 +1,10 @@
+<?php
+
+/** @deprecated */
+abstract class tx_rnbase_tests_BaseTestCase extends Sys25\RnBase\Tests\BaseTestCase
+{
+}
+/** @deprecated */
+class tx_rnbase_tests_Utility extends Sys25\RnBase\Tests\TestUtility
+{
+}